backend/requirements.txt
# backend/requirements.txt

fastapi
uvicorn[standard]
httpx
redis[hiredis]>=4.4.0 # Use a version supporting async and hiredis for performance
python-dotenv         # For loading .env file (optional but good practice)
pydantic              # Explicitly list for clarity, though fastapi depends on it
pydantic_settings
